+++ Commands/Model/Footer.cs
@@ -1,0 +1,13 @@
+﻿namespace SharePointPnP.PowerShell.Commands.Model
+{
+    /// <summary>
+    /// Contains the properties for working with the footer on a web
+    /// </summary>
+    public class Footer
+    {
+        /// <summary>
+        /// Boolean indicating if the footer is enabled
+        /// </summary>
+        public bool? IsEnabled;
+    }
+}